centralise snake helpers :snake:
